unified_trading_platform/requirements.txt
# Core Dependencies
asyncio-mqtt==0.13.0
aiohttp==3.8.5
aiofiles==23.2.1
pydantic==2.3.0
python-dotenv==1.0.0
PyYAML==6.0.1
click==8.1.7
rich==13.5.2
typer==0.9.0

# Data Processing & Analysis
pandas==2.0.3
numpy==1.24.3
scipy==1.11.1
scikit-learn==1.3.0
statsmodels==0.14.0

# AI & Machine Learning
tensorflow==2.13.0
torch==2.0.1
transformers==4.32.1
joblib==1.3.2
xgboost==1.7.6
lightgbm==4.0.0

# Financial Data & Trading
ccxt==4.0.77
yfinance==0.2.18
alpha-vantage==2.3.1
websocket-client==1.6.1
python-binance==1.0.19

# Database & Storage
SQLAlchemy==2.0.20
alembic==1.11.3
redis==4.6.0
motor==3.3.1  # Async MongoDB
psycopg2-binary==2.9.7  # PostgreSQL

# Web Framework & Dashboard
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
Jinja2==3.1.2
Werkzeug==2.3.7
eventlet==0.33.3
python-socketio==5.8.0

# API & HTTP
requests==2.31.0
httpx==0.24.1
fastapi==0.103.1
uvicorn==0.23.2
starlette==0.27.0

# Monitoring & Performance
psutil==5.9.5
prometheus-client==0.17.1
py-spy==0.3.14
memory-profiler==0.60.0
line-profiler==4.1.1

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest-benchmark==4.0.0
hypothesis==6.82.6

# Logging & Configuration
structlog==23.1.0
loguru==0.7.0
colorlog==6.7.0
python-json-logger==2.0.7

# Security & Authentication
cryptography==41.0.4
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0
APScheduler==3.10.4
cachetools==5.3.1

# Development & Code Quality
black==23.7.0
flake8==6.0.0
mypy==1.5.1
isort==5.12.0
pre-commit==3.3.3

# Documentation
Sphinx==7.1.2
sphinx-rtd-theme==1.3.0

# Optional: Advanced Features
# Uncomment as needed

# GPU Computing (for advanced AI models)
# cupy-cuda11x==12.2.0

# Distributed Computing
# dask==2023.8.1
# ray==2.6.3

# Advanced Visualization
# plotly==5.15.0
# bokeh==3.2.2
# dash==2.13.0

# Message Queues
# celery==5.3.1
# kombu==5.3.1
# pika==1.3.2  # RabbitMQ

# Time Series Databases
# influxdb-client==1.37.0
# pymongo==4.4.1

# Advanced Analytics
# ta-lib==0.4.25  # Technical Analysis Library
# zipline==1.4.1  # Algorithmic Trading
# backtrader==1.9.78.123  # Backtesting

# Cloud Services
# boto3==1.28.25  # AWS
# google-cloud-storage==2.10.0  # GCP
# azure-storage-blob==12.17.0  # Azure

# Blockchain & Crypto
# web3==6.9.0
# eth-account==0.9.0
# bitcoin==1.1.42 